Refactor Filament Widgets and Enhance Chart Configuration
- Moved `WorkingMinutesChartWidget.php`, `WorkHoursChartWidget.php`, `ParticipationInvoiceChartWidget.php`, and `IssuedInvoiceChartWidget.php` under `Charts` namespace for better organization.
- Updated import paths in `AgentStatsDashboard.php` to reflect the namespace changes.
- Added grid display options to `WorkingMinutesChartWidget`, `WorkHoursChartWidget`, `ParticipationInvoiceChartWidget`, and `IssuedInvoiceChartWidget` to enhance chart configurations.
- Removed unused `paginateTableQuery` method and unnecessary actions from `InvoicesWidget`.

These changes improve code organization and enhance the display options for charts in Filament dashboards and widgets.
